Add Thread, Coroutine, Port and Channel to dart:core.
Until we find a library for these members, we expose them by taking all
non-@patch in core_patch, and add it to the scope of dart:core.

BUG=
R=zerny@google.com

Review URL: https://codereview.chromium.org/1068763002

+++ lib/simple_system/double.dart
@@ -187,7 +187,7 @@ class double implements core.double {
 
   double _divFromInteger(int other) => other.toDouble() / this;
 
-  double _truncDivFromInteger(int other) => other.toDouble() ~/ this;
+  int _truncDivFromInteger(int other) => other.toDouble() ~/ this;
 
   bool _compareEqFromInteger(int other) => other.toDouble() == this;
